--- examples/next/docs/js/accessibility/constants.js
@@ -1,2 +1,0 @@
-export const SELECTED_CLASS = "selected";
-export const ODD_ROW_CLASS = "odd";
