+++ examples/next/docs/js/accessibility/data.js
@@ -14,37 +14,28 @@ export const data = [
   {
     companyName: randCompanyName({ maxCharCount: 18 }),
     productName: randProductName({ maxCharCount: 18 }),
-    sellDate: randPastDate().toLocaleDateString('en-gb'),
+    sellDate: randPastDate().toLocaleDateString("en-gb"),
     inStock: randBoolean(),
     qty: randNumber({ max: 500 }),
-    progress: randNumber({ max: 10 }),
-    rating: randNumber({ max: 5 }),
-    orderId: `${randNumber({ max: 9, min: 0 })}${randNumber({ max: 9, min: 0 })}-${randNumber({
+    orderId: `${randNumber({ max: 9, min: 0 })}${randNumber({
+      max: 9,
+      min: 0,
+    })}-${randNumber({
       max: 9999999,
       min: 1111111,
     })}`,
-    country: 'United Kingdom',
-    __children: [
-      {
-        companyName: randCompanyName({ maxCharCount: 18 }),
-      },
-      {
-        companyName: randCompanyName({ maxCharCount: 18 }),
-      },
-      {
-        companyName: randCompanyName({ maxCharCount: 18 }),
-      },
-    ],
+    country: "United Kingdom",
   },
   ...Array.from({ length: 150 }, () => ({
     companyName: randCompanyName({ maxCharCount: 18 }),
     productName: randProductName({ maxCharCount: 18 }),
-    sellDate: randPastDate().toLocaleDateString('en-gb'),
+    sellDate: randPastDate().toLocaleDateString("en-gb"),
     inStock: randBoolean(),
     qty: randNumber({ max: 500 }),
-    progress: randNumber({ max: 10 }),
-    rating: randNumber({ max: 5 }),
-    orderId: `${randNumber({ max: 9, min: 0 })}${randNumber({ max: 9, min: 0 })}-${randNumber({
+    orderId: `${randNumber({ max: 9, min: 0 })}${randNumber({
+      max: 9,
+      min: 0,
+    })}-${randNumber({
       max: 9999999,
       min: 1111111,
     })}`,
