--- examples/next/docs/react/accessibility/src/constants.ts
@@ -1,3 +1,0 @@
-
-export const SELECTED_CLASS = "selected";
-export const ODD_ROW_CLASS = "odd";
