+++ examples/next/docs/react/accessibility/src/hooksCallbacks.ts
@@ -1,45 +1,10 @@
 import Handsontable from "handsontable";
-import { ODD_ROW_CLASS } from "./constants";
 
 const headerAlignments = new Map([
   ["9", "htCenter"],
   ["10", "htRight"],
   ["12", "htCenter"],
 ]);
-
-type AddClassesToRows = (
-  TD: HTMLTableCellElement,
-  row: number,
-  column: number,
-  prop: number | string,
-  value: any,
-  cellProperties: Handsontable.CellProperties
-) => void;
-
-export const addClassesToRows: AddClassesToRows = (
-  TD,
-  row,
-  column,
-  _prop,
-  _value
-) => {
-  // Adding classes to `TR` just while rendering first visible `TD` element
-  if (column !== 0) {
-    return;
-  }
-
-  const parentElement = TD.parentElement;
-
-  if (parentElement === null) {
-    return;
-  }
-  // Add class to odd TRs
-  if (row % 2 === 0) {
-    Handsontable.dom.addClass(parentElement, ODD_ROW_CLASS);
-  } else {
-    Handsontable.dom.removeClass(parentElement, ODD_ROW_CLASS);
-  }
-};
 
 export function alignHeaders(
   this: Handsontable,
